Move the `promise-util` module under the `@bayou` namespace.
