+++ local-modules/@bayou/promise-util/EventSource.js
