+++ local-modules/api-server/package.json
@@ -8,7 +8,7 @@
     "api-common": "local",
     "@bayou/codec": "local",
     "hooks-server": "local",
-    "promise-util": "local",
+    "@bayou/promise-util": "local",
     "@bayou/see-all": "local",
     "@bayou/typecheck": "local",
     "@bayou/util-common": "local"
