+++ local-modules/dev-mode/DevMode.js
@@ -8,7 +8,7 @@ import fs_extra from 'fs-extra';
 import path from 'path';
 
 import { Dirs } from 'env-server';
-import { Delay } from 'promise-util';
+import { Delay } from '@bayou/promise-util';
 import { Logger } from '@bayou/see-all';
 import { Singleton } from '@bayou/util-common';
