+++ local-modules/dev-mode/package.json
@@ -7,7 +7,7 @@
     "fs-extra": "^3.0.1",
 
     "env-server": "local",
-    "promise-util": "local",
+    "@bayou/promise-util": "local",
     "@bayou/see-all": "local",
     "@bayou/util-common": "local"
   }
