+++ local-modules/doc-client/BodyClient.js
@@ -4,7 +4,7 @@
 
 import { ConnectionError } from 'api-common';
 import { BodyChange, BodyDelta, BodySnapshot } from 'doc-common';
-import { Delay } from 'promise-util';
+import { Delay } from '@bayou/promise-util';
 import { QuillEvents, QuillUtil } from 'quill-util';
 import { TString } from '@bayou/typecheck';
 import { StateMachine } from 'state-machine';
