+++ local-modules/doc-server/BaseControl.js
@@ -5,7 +5,7 @@
 import { Timeouts } from 'doc-common';
 import { Errors as fileStoreOt_Errors, StoragePath, TransactionSpec } from 'file-store-ot';
 import { BaseSnapshot, RevisionNumber } from 'ot-common';
-import { Delay } from 'promise-util';
+import { Delay } from '@bayou/promise-util';
 import { TBoolean, TFunction } from '@bayou/typecheck';
 import { Errors } from '@bayou/util-common';
