+++ local-modules/doc-server/FileBootstrap.js
@@ -6,7 +6,7 @@ import { BodyChange, BodyDelta } from 'doc-common';
 import { TransactionSpec } from 'file-store-ot';
 import { DEFAULT_DOCUMENT } from 'hooks-server';
 import { Timestamp } from 'ot-common';
-import { Mutex } from 'promise-util';
+import { Mutex } from '@bayou/promise-util';
 import { Errors } from '@bayou/util-common';
 
 import BaseDataManager from './BaseDataManager';
