+++ local-modules/doc-server/package.json
@@ -13,7 +13,7 @@
     "file-store-ot": "local",
     "hooks-server": "local",
     "ot-common": "local",
-    "promise-util": "local",
+    "@bayou/promise-util": "local",
     "@bayou/see-all": "local",
     "@bayou/typecheck": "local",
     "@bayou/util-common": "local"
