+++ local-modules/file-store-local/LocalFile.js
@@ -9,7 +9,7 @@ import { Codec } from '@bayou/codec';
 import { BaseFile } from 'file-store';
 import { FileChange, FileSnapshot } from 'file-store-ot';
 import { RevisionNumber } from 'ot-common';
-import { Condition, Delay, Mutex } from 'promise-util';
+import { Condition, Delay, Mutex } from '@bayou/promise-util';
 import { TString } from '@bayou/typecheck';
 import { Logger } from '@bayou/see-all';
 import { FrozenBuffer, Errors } from '@bayou/util-common';
