+++ local-modules/file-store/package.json
@@ -8,7 +8,7 @@
     "@bayou/codec": "local",
     "file-store-ot": "local",
     "hooks-server": "local",
-    "promise-util": "local",
+    "@bayou/promise-util": "local",
     "@bayou/typecheck": "local",
     "@bayou/util-common": "local"
   }
