+++ local-modules/hooks-server/package.json
@@ -6,7 +6,7 @@
     "api-server": "local",
     "file-store-local": "local",
     "hooks-common": "local",
-    "promise-util": "local",
+    "@bayou/promise-util": "local",
     "@bayou/util-common": "local"
   }
 }
