+++ local-modules/quill-util/QuillProm.js
@@ -4,7 +4,7 @@
 
 import Quill from 'quill';
 
-import { EventSource } from 'promise-util';
+import { EventSource } from '@bayou/promise-util';
 import { TObject } from '@bayou/typecheck';
 import { CommonBase, Functor } from '@bayou/util-common';
