+++ local-modules/quill-util/package.json
@@ -5,7 +5,7 @@
   "dependencies": {
     "@bayou/deps-quill-client": "local",
     "doc-common": "local",
-    "promise-util": "local",
+    "@bayou/promise-util": "local",
     "@bayou/typecheck": "local",
     "@bayou/util-common": "local"
   }
