+++ local-modules/testing-client/package.json
@@ -5,7 +5,7 @@
   "dependencies": {
     "@bayou/deps-testing": "local",
     "@bayou/see-all": "local",
-    "promise-util": "local",
+    "@bayou/promise-util": "local",
     "mocha-client-shim": "local",
     "testing-common": "local",
     "@bayou/util-common": "local"
